backend/backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
stripe
gunicorn
whitenoise